chore: add tracking for CLI relatd to creator hub
